# blit_impl: Use <algorithm> with unseq policy