On my desktop with Ryzen 3950x, RTX 4080, Ubuntu 24.04, gcc 13.2.0, I get a consistent performance improvement with this PR, as measured by `tools/measure_timedemo_performance.py`.

Before: 7.722 ± 0.065 seconds, 949.090 ± 7.807 FPS\
After: 6.733 ± 0.057 seconds, 1088.520 ± 8.710 FPS

Binary size is also decreased by ~350 KiB:

Before: 6007752
After: 5652488

RG99 binary size is only decreased by ~16 KiB, still a good thing.