@AJenbo 2024-06-09
Fixed my cooling solution and upgraded to 24.04 and now I'm also seeing an improvement:
Ryzen 5950x

Before: 1380fps
After: 1610fps